LLMDriverAssistant/requirements.txt
 langchain>=0.1.16     
 langchain-openai>=0.1.0
 langchain-chroma>=0.1.0
 chromadb==0.6.3
 openai>=1.14.3
 tiktoken>=0.6.0
 confluent-kafka>=2.4.0
 python-dotenv>=1.0.1
